Add files via upload
This is a fun project done by my self with the reference of Tech with Tim. It is created using Turtle modules provided by python which race between a range of turtles from 2 to 10 and announce the winner at the end of the race with the colour tag of the turtle. The purpose is to understand more fundamentals of python on graphics and gaming.